Changed to use boost:shared_ptr.

+++ ModFossaCpp/test/VoltageProtocolTest.cpp
@@ -47,7 +47,7 @@ TEST_F(VoltageProtocolTest, addConstantStageSuccess) {
     vp = new VoltageProtocol("vp1");
     vp->addConstantStage("stage1", -50, 200);
     
-    ASSERT_EQ(vp->serializeVoltageProtocol()->size(), 1);
+    ASSERT_EQ(vp->serializeVoltageProtocol()->size(), (unsigned int)1);
 }
 
 /**
